// utils/types.ts
  export type UserCredentialsParams = {
    username: string;
    password: string;
  };